training/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.23
Pillow>=10.0.0
opencv-python-headless>=4.8.0.74
tqdm>=4.66.0
scikit-image>=0.22.0
tyro>=0.7.0